requirements.txt
BeautifulSoup4
lxml
click

langchain
openai
python-dotenv